# bank-backend/api-calls/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/transactions/{id}/dispute": {
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "raiseDispute",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RaiseDisputeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{id}/cancel": {
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "cancelTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/withdraw": {
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "withdraw",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/search": {
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "searchTransactions",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchTransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/export": {
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "exportTransactions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportTransactionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/deposit": {
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "deposit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepositRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/transfers/{id}/receipt": {
      "get": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "getReceipt",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferReceiptResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/statistics": {
      "get": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "getStatistics",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferStatisticsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/transactions": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "getAllTransactions",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "minAmount",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxAmount",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{id}": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "getTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{id}/receipt": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "getReceipt_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionReceiptResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/statistics": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "getStatistics_1",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionStatisticsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/recent": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "getRecentTransactions",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/pending": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "getPendingTransactions",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTransactionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/payments/{id}/receipt": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "getReceipt_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentReceiptResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/statistics": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "getStatistics_2",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentStatisticsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cards/statistics": {
      "get": {
        "tags": [
          "card-controller"
        ],
        "operationId": "getStatistics_3",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardStatisticsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "RaiseDisputeRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": [
          "reason"
        ]
      },
      "WithdrawRequest": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number",
            "minimum": 0.01
          },
          "description": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": [
          "accountId",
          "amount"
        ]
      },
      "TransactionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "accountId": {
            "type": "integer",
            "format": "int64"
          },
          "transactionType": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SearchTransactionRequest": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "minAmount": {
            "type": "number"
          },
          "maxAmount": {
            "type": "number"
          }
        }
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PageTransactionResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "ExportTransactionsRequest": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "format": {
            "type": "string",
            "minLength": 1,
            "pattern": "pdf|csv|PDF|CSV"
          },
          "accountId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "endDate",
          "format",
          "startDate"
        ]
      },
      "DepositRequest": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number",
            "minimum": 0.01
          },
          "description": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": [
          "accountId",
          "amount"
        ]
      },
...
      "TransferReceiptResponse": {
        "type": "object",
        "properties": {
          "transferId": {
            "type": "integer",
            "format": "int64"
          },
          "reference": {
            "type": "string"
          },
          "transferType": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "fromAccountNumber": {
            "type": "string"
          },
          "toAccountNumber": {
            "type": "string"
          },
          "recipientName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "transferDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TransferStatisticsResponse": {
        "type": "object",
        "properties": {
          "totalTransfers": {
            "type": "integer",
            "format": "int32"
          },
          "internalTransfers": {
            "type": "integer",
            "format": "int32"
          },
          "externalTransfers": {
            "type": "integer",
            "format": "int32"
          },
          "totalAmount": {
            "type": "number"
          },
          "completedTransfers": {
            "type": "integer",
            "format": "int32"
          },
          "pendingTransfers": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "TransactionReceiptResponse": {
        "type": "object",
        "properties": {
          "transactionId": {
            "type": "integer",
            "format": "int64"
          },
          "reference": {
            "type": "string"
          },
          "transactionType": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "accountNumber": {
            "type": "string"
          },
          "accountType": {
            "type": "string"
          },
          "transactionDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TransactionStatisticsResponse": {
        "type": "object",
        "properties": {
          "totalTransactions": {
            "type": "integer",
            "format": "int32"
          },
          "totalDeposits": {
            "type": "number"
          },
          "totalWithdrawals": {
            "type": "number"
          },
          "depositCount": {
            "type": "integer",
            "format": "int32"
          },
          "withdrawalCount": {
            "type": "integer",
            "format": "int32"
          },
          "pendingCount": {
            "type": "integer",
            "format": "int32"
          },
          "failedCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PaymentReceiptResponse": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "integer",
            "format": "int64"
          },
          "reference": {
            "type": "string"
          },
          "paymentType": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "payeeName": {
            "type": "string"
          },
          "payeeAccount": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "fromAccountNumber": {
            "type": "string"
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PaymentStatisticsResponse": {
        "type": "object",
        "properties": {
          "totalPayments": {
            "type": "integer",
            "format": "int32"
          },
          "billPayments": {
            "type": "integer",
            "format": "int32"
          },
          "utilityPayments": {
            "type": "integer",
            "format": "int32"
          },
          "merchantPayments": {
            "type": "integer",
            "format": "int32"
          },
          "totalAmount": {
            "type": "number"
          },
          "completedPayments": {
            "type": "integer",
            "format": "int32"
          },
          "pendingPayments": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}